// ecomerce-frontend/src/app/admin/interfaces/reportes.interface.ts
export interface Reporte{

  cantidadVentas?:      number;
  nombreProducto?:      string;
  cantidadProductos?:   number;
  nombreUsuario?:       string;
  ganancias?:           number;
  cantidadPedidos?:     number;
  nombre?:              string;
  cliente?:             string;
  usuario?:             string;
  totalProductos?:      string;
  totalCompras?:        number;
  count?:               number;
}

export interface RangoFechas{
  fechaInicio:         Date;
  fechaFin:            Date;
}